atof done, todo get_vector and parent functions

+++ includes/vector.h
@@ -6,7 +6,7 @@
 /*   By: darbib <darbib@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/02/28 15:03:37 by darbib            #+#    #+#             */
-/*   Updated: 2020/02/28 15:05:26 by darbib           ###   ########.fr       */
+/*   Updated: 2020/03/02 18:47:46 by darbib           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@
 
 typedef struct		s_vect
 {
-	float			x;
-	float			y;
-	float			z;
+	double			x;
+	double			y;
+	double			z;
 }					t_vect;
 
 #endif
